refactor(scheduling): 重构排课相关逻辑和数据结构
- 修改了多个类的包结构，优化了项目组织
- 重构了部分类的名称，使其更具描述性
- 调整了部分方法的签名，提高了代码可读性
- 新增了教学班相关实体和逻辑
- 优化了课程类型获取逻辑- 改进了初始种群生成方式

+++ src/main/java/com/frontleaves/scheduling/models/dto/scheduling/ScheduleResultDTO.java
@@ -102,7 +102,7 @@ public class ScheduleResultDTO {
         /**
          * 优先级
          */
-        private Short priority;
+        private Integer priority;
     }
 
     @Data
